backend/requirements.txt
Flask==2.0.1
flask-cors==3.0.10
twilio==7.16.0
python-dotenv==0.19.1
Pillow==9.4.0
pytesseract==0.3.10
PyMuPDF==1.21.1
google-generativeai==0.2.0
Werkzeug==2.0.1
